refactor(arch): Phase 4 — PR #485 리뷰 수정
- B1: 중복 import 7곳 병합 (@/entities/ticker, @/entities/analysis)
- B2: shared/lib/types.ts → entities 상향 의존 해소 (타입 직접 정의)
- B3: shared/db/ eslint-disable 제거, LLM_PROVIDER_VALUES → shared/config/llmProviders.ts
- S1: queryConfig.ts 상대경로 통일
