+++ src/app/[symbol]/fear-greed/page.tsx
@@ -2,10 +2,12 @@ import { FearGreedPage } from '@/components/fear-greed/FearGreedPage';
 import { CrossLinkCards } from '@/components/symbol-page/CrossLinkCards';
 import { JsonLd } from '@/shared/ui/JsonLd';
 import { DEFAULT_TIMEFRAME, VALID_TICKER_RE } from '@/shared/config/market';
-import { buildAssetAboutNode } from '@/entities/ticker';
-import { buildDisplayName } from '@/entities/ticker';
+import {
+    buildAssetAboutNode,
+    buildDisplayName,
+    getAssetInfoCached,
+} from '@/entities/ticker';
 import { getBarsAction } from '@/entities/bars/actions';
-import { getAssetInfoCached } from '@/entities/ticker';
 import { QUERY_KEYS, QUERY_STALE_TIME_MS } from '@/shared/config/queryConfig';
 import {
     buildBreadcrumbJsonLd,
