+++ src/app/[symbol]/news/page.tsx
@@ -13,11 +13,13 @@ import { CrossLinkCards } from '@/components/symbol-page/CrossLinkCards';
 import { SectionSkeleton } from '@/components/symbol-page/SectionSkeleton';
 import { JsonLd } from '@/shared/ui/JsonLd';
 import { VALID_TICKER_RE } from '@/shared/config/market';
-import { buildAssetAboutNode } from '@/entities/ticker';
-import { buildDisplayName } from '@/entities/ticker';
+import {
+    buildAssetAboutNode,
+    buildDisplayName,
+    getAssetInfoCached,
+} from '@/entities/ticker';
 import { ensureNewsCardsAnalyzedAction } from '@/entities/news-article/actions';
 import { getTodayIsoDay } from '@/shared/lib/getTodayIsoDay';
-import { getAssetInfoCached } from '@/entities/ticker';
 import { todayKstIsoDate } from '@/shared/lib/dateKey';
 import {
     buildBreadcrumbJsonLd,
