+++ src/app/[symbol]/overall/page.tsx
@@ -6,9 +6,11 @@ import {
     isValidTimeframe,
     VALID_TICKER_RE,
 } from '@/shared/config/market';
-import { buildAssetAboutNode } from '@/entities/ticker';
-import { buildDisplayName } from '@/entities/ticker';
-import { getAssetInfoCached } from '@/entities/ticker';
+import {
+    buildAssetAboutNode,
+    buildDisplayName,
+    getAssetInfoCached,
+} from '@/entities/ticker';
 import {
     buildBreadcrumbJsonLd,
     buildSymbolOverallSeoContent,
